+++ plugins/src/render.gravity_lines/main.cpp
@@ -408,7 +408,7 @@ class vsx_module_gravity_ribbon_mesh : public vsx_module {
 
   vsx_matrix modelview_matrix_no_connection;
   float last_update;
-  long prev_num_vertices;
+  unsigned long prev_num_vertices;
 
 public:
   
@@ -509,10 +509,10 @@ public:
 
     if (mesh && (*mesh)->data->vertices.size())
     {
-      if (prev_num_vertices != (int)mesh_id_count->get())
+      if (prev_num_vertices != (unsigned long)mesh_id_count->get())
       {
         // remove all the old ones
-        for (unsigned long i = prev_num_vertices; i < (int)mesh_id_count->get(); ++i)
+        for (unsigned long i = prev_num_vertices; i < (unsigned long)mesh_id_count->get(); ++i)
         {
           //printf("allocating %d\n", i);
           //if (i == prev_num_vertices)
@@ -624,7 +624,7 @@ public:
         upv.x = upvector->get(0);
         upv.y = upvector->get(1);
         upv.z = upvector->get(2);
-        for (int i = 0; i < prev_num_vertices; ++i)
+        for (unsigned long i = 0; i < prev_num_vertices; ++i)
         {
           gr[i]->width = ribbon_width->get();
           gr[i]->length = length->get();
